// src/core/interfaces/registrationObjectInterface.ts
export interface IRegistrationObject {
    email: string;
    password: string;
    name: string;
    lastName: string;
    country: string;
    city: string;
    street: string;
    postcode: string;
    birthDate: string;
    countryShip: string;
    cityShip: string;
    streetShip: string;
    postcodeShip: string;
    billingDefault: boolean;
    shippingDefault: boolean;
}